+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Added
   - Drag and drop from netlist view to add variables into the viewer
   - Signal height can now be set to 2x, 4x, and 8x
+  - Analog signals can now be vertically zoomed in and out
   - ASCII value format
 - API
   - Added (proposed) onDidDropInWaveformViewer event